fix: Make uploader template wrapper
